+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 <!-- ## [Unreleased] -->
 
+## [0.3.2] - 2022-03-24
+
+### Added
+
+- `--verbose` prints full diff as reported by `tcompare`
+
+### Fixed
+
+- actual vs expected colors in equality assertions
+
 ## [0.3.1] - 2022-03-24
 
 ### Changed